It works! Downloader added
After extensive mucking around with the project.json frameworks and
dependencies, I was finally able to get the project to recognize all the
methods and variables of both packages.

Now that I have the functionality, I gotta test it at home. There's also
plenty of error control and deterence to program in.